requirements.txt
selenium==4.18.1
pytest==8.0.2
allure-pytest==2.13.2
requests==2.31.0
python-dotenv==1.0.1
